requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
shap
scipy
requests
beautifulsoup4
lxml
openpyxl
pyarrow
fastparquet
jinja2
python-dotenv
anthropic
openai
gradio
